+++ discord/channel.py
@@ -320,6 +320,13 @@ class _TextChannel(discord.abc.GuildChannel, Hashable):
         """Checks if the channel is NSFW."""
         return self.nsfw
 
+    def is_spoiler(self) -> bool:
+        """Checks if the channel is a spoiler channel.
+
+        .. versionadded:: 2.9
+        """
+        return self.flags.is_spoiler_channel
+
     @property
     def last_message(self) -> Message | None:
         """Fetches the last message from this channel in cache.
@@ -812,6 +819,7 @@ class TextChannel(discord.abc.Messageable, _TextChannel):
         default_thread_slowmode_delay: int = ...,
         type: ChannelType = ...,
         overwrites: Mapping[Role | Member | Snowflake, PermissionOverwrite] = ...,
+        spoiler: bool = ...,
     ) -> TextChannel | None: ...
 
     @overload
@@ -868,6 +876,10 @@ class TextChannel(discord.abc.Messageable, _TextChannel):
             The new default slowmode delay in seconds for threads created in this channel.
 
             .. versionadded:: 2.3
+        spoiler: :class:`bool`
+            Whether the channel should be a spoiler channel. Mutually exclusive with :attr:`nsfw`.
+
+            .. versionadded:: 2.9
 
         Returns
         -------
@@ -885,6 +897,10 @@ class TextChannel(discord.abc.Messageable, _TextChannel):
         HTTPException
             Editing the channel failed.
         """
+        if "spoiler" in options:
+            options["flags"] = ChannelFlags._from_value(self.flags.value)
+            options["flags"].is_spoiler_channel = options.pop("spoiler")
+
         payload = await self._edit(options, reason=reason)
         if payload is not None:
             # the payload will always be the proper channel payload
@@ -1126,6 +1142,7 @@ class ForumChannel(_TextChannel):
         available_tags: list[ForumTag] = ...,
         require_tag: bool = ...,
         overwrites: Mapping[Role | Member | Snowflake, PermissionOverwrite] = ...,
+        spoiler: bool = ...,
     ) -> ForumChannel | None: ...
 
     @overload
@@ -1188,6 +1205,10 @@ class ForumChannel(_TextChannel):
             Whether a tag should be required to be specified when creating a thread in this channel.
 
             .. versionadded:: 2.3
+        spoiler: :class:`bool`
+            Whether the channel should be a spoiler channel. Mutually exclusive with :attr:`nsfw`.
+
+            .. versionadded:: 2.9
 
         Returns
         -------
@@ -1208,6 +1229,10 @@ class ForumChannel(_TextChannel):
         if "require_tag" in options:
             options["flags"] = ChannelFlags._from_value(self.flags.value)
             options["flags"].require_tag = options.pop("require_tag")
+        if "spoiler" in options:
+            if "flags" not in options:
+                options["flags"] = ChannelFlags._from_value(self.flags.value)
+            options["flags"].is_spoiler_channel = options.pop("spoiler")
 
         payload = await self._edit(options, reason=reason)
         if payload is not None:
@@ -1507,6 +1532,7 @@ class MediaChannel(ForumChannel):
         require_tag: bool = ...,
         hide_media_download_options: bool = ...,
         overwrites: Mapping[Role | Member | Snowflake, PermissionOverwrite] = ...,
+        spoiler: bool = ...,
     ) -> ForumChannel | None: ...
 
     async def edit(self, *, reason=None, **options):
@@ -1563,6 +1589,11 @@ class MediaChannel(ForumChannel):
         hide_media_download_options: :class:`bool`
             Whether media download options should be hidden in this media channel.
 
+        spoiler: :class:`bool`
+            Whether the channel should be a spoiler channel. Mutually exclusive with :attr:`nsfw`.
+
+            .. versionadded:: 2.9
+
         Returns
         -------
         Optional[:class:`.MediaChannel`]
@@ -1580,14 +1611,18 @@ class MediaChannel(ForumChannel):
             Editing the channel failed.
         """
 
-        if "require_tag" in options or "hide_media_download_options" in options:
+        if (
+            "require_tag" in options
+            or "hide_media_download_options" in options
+            or "spoiler" in options
+        ):
             flags = ChannelFlags._from_value(self.flags.value)
             flags.require_tag = options.pop("require_tag", flags.require_tag)
             flags.hide_media_download_options = options.pop(
                 "hide_media_download_options", flags.hide_media_download_options
             )
+            flags.is_spoiler_channel = options.pop("spoiler", flags.is_spoiler_channel)
             options["flags"] = flags
-
         payload = await self._edit(options, reason=reason)
         if payload is not None:
             # the payload will always be the proper channel payload
@@ -1817,6 +1852,13 @@ class VoiceChannel(discord.abc.Messageable, VocalGuildChannel):
     def is_nsfw(self) -> bool:
         """Checks if the channel is NSFW."""
         return self.nsfw
+
+    def is_spoiler(self) -> bool:
+        """Checks if the channel is a spoiler channel.
+
+        .. versionadded:: 2.9
+        """
+        return self.flags.is_spoiler_channel
 
     @property
     def last_message(self) -> Message | None:
@@ -2099,6 +2141,7 @@ class VoiceChannel(discord.abc.Messageable, VocalGuildChannel):
         slowmode_delay: int = ...,
         nsfw: bool = ...,
         reason: str | None = ...,
+        spoiler: bool = ...,
     ) -> VoiceChannel | None: ...
 
     @overload
@@ -2157,6 +2200,11 @@ class VoiceChannel(discord.abc.Messageable, VocalGuildChannel):
 
             .. versionadded:: 2.7
 
+        spoiler: :class:`bool`
+            Whether the channel should be a spoiler channel. Mutually exclusive with :attr:`nsfw`.
+
+            .. versionadded:: 2.9
+
         Returns
         -------
         Optional[:class:`.VoiceChannel`]
@@ -2172,6 +2220,9 @@ class VoiceChannel(discord.abc.Messageable, VocalGuildChannel):
         HTTPException
             Editing the channel failed.
         """
+        if "spoiler" in options:
+            options["flags"] = ChannelFlags._from_value(self.flags.value)
+            options["flags"].is_spoiler_channel = options.pop("spoiler")
 
         payload = await self._edit(options, reason=reason)
         if payload is not None:
@@ -2405,6 +2456,13 @@ class StageChannel(discord.abc.Messageable, VocalGuildChannel):
     def is_nsfw(self) -> bool:
         """Checks if the channel is NSFW."""
         return self.nsfw
+
+    def is_spoiler(self) -> bool:
+        """Checks if the channel is a spoiler channel.
+
+        .. versionadded:: 2.9
+        """
+        return self.flags.is_spoiler_channel
 
     @property
     def last_message(self) -> Message | None: