+++ CHANGELOG.md
@@ -12,6 +12,11 @@ These changes are available on the `master` branch, but have not yet been releas
 
 ### Added
 
+- Added `ChannelFlags.is_spoiler_channel` and `.is_spoiler()` function to all applicable
+  channel types. ([#3252](https://github.com/Pycord-Development/pycord/pull/3252))
+- Added `spoiler` parameter to the `edit()` function of all applicable channel types.
+  ([#3252](https://github.com/Pycord-Development/pycord/pull/3252))
+
 ### Changed
 
 ### Fixed